Fix decimal scores and grading-mode reset on mobile
Điểm lẻ (7.5) bị mất: màn xác nhận Unit Test parse điểm bằng int.tryParse
nên "7.5" ra null rồi lưu thành 0, bàn phím số lại không có phím thập phân
để gõ, backend validate sub_scores là integer, và schema/prompt bắt AI trả
INTEGER nên nửa điểm trên bài bị làm tròn ngay từ bước đọc ảnh. Giờ điểm
thành phần và TỔNG đi kiểu float suốt chuỗi (AI -> API -> app), ô nhập bật
bàn phím thập phân và nhận cả "7,5", so khớp tổng dùng sai số thay vì !=.

Đề đang chấm tự nhảy về kiểu "đếm câu đúng": getClassExam() nuốt mọi lỗi
thành null, nên chỉ cần một nhịp mất mạng/hết token là app tưởng lớp chưa
có đề, mở lại hộp thoại kiểu chấm (mặc định đếm câu đúng) rồi POST
/exams/today ghi đè grading_mode của đề đang chấm dở — hỏng vĩnh viễn.
Giờ chỉ nuốt 404, hộp thoại tạo đề mới điền sẵn kiểu chấm của đề gần nhất,
và màn chụp có nút chuyển thẳng sang luồng 2 ảnh để không bị kẹt ở option 1
(trước đó lối thoát duy nhất là tạo đề mới, mà tạo đề mới thì khoá luôn đề
đang chấm dở).

+++ backend/app/Services/Vision/GradedPaperResult.php
@@ -5,11 +5,14 @@ namespace App\Services\Vision;
 class GradedPaperResult
 {
     /**
-     * @param  array{vocabulary:?int,grammar:?int,listening:?int,reading:?int,writing:?int,speaking:?int}|null  $subScores
+     * Điểm để kiểu float (không phải int) vì giáo viên chấm tay thường ghi
+     * điểm lẻ 0.5 trên bài — đọc thành int là mất luôn phần lẻ đó.
+     *
+     * @param  array{vocabulary:?float,grammar:?float,listening:?float,reading:?float,writing:?float,speaking:?float}|null  $subScores
      */
     public function __construct(
         public readonly ?string $studentName,
-        public readonly ?int $totalScore,
+        public readonly ?float $totalScore,
         public readonly ?array $subScores,
         public readonly float $confidence,
     ) {}
